Fix Chande-Kroll stop #256, and #257
Add an optional "tvmode" parameter that controls the behaviour of the
Chande-Kroll volatility stop. The default, True, retains the used
behaviour, compatibility with Trading View. When False however, the
moving average mode used is now a simple moving average instead of the
Welles Wilder moving average and the periods used are different as well,
(10/1/9 vs 10/3/20).

Update the unit tests to account for the optional parameter, though
sadly we have no TA-lib implementation to compare with, and the Japanese
Yen Futures 09/93 contract used in figure 7.4 page 95 of the book, has
no freely available OHLC data i could find.