Fix 'verbose' option in multiprocessing mode
The verbose option does not take effect multiprocessing mode. See L746 and L751 in core.py.
The "without multiprocessing" part deals with this option correctly (L772 and L782).

This may not be an issue when use in jupyter notebook, but it's a bit annoying when running in terminal, especially within for loop.