Add initial CRSI indicator
Rely on Pandas-TA built-in Wilder's RSI indicator and TA-lib
counterparts, and implement percent_rank and streak calculation
functions using NumPy. This was tested against large time series with 1
million entries and found to be performant as detailed in the PR ticket.
Add separate streak and percent rank functions with doctests.