Add fix for #742
Add sum_signed_rolling_deltas function and proper NumPy vectorization.
Some TMO indicators in TV consider the rolling window to be exclusive
of the current day, while others consider it inclusive. Since this
series is used to compute the moving averages and momentums for the
TMO, there would be a plot mismatch against TV unless we dealt with
both cases.
Added doctests for both cases, noting that since the series must be cast
to float type, fp precision representation issues might arise, which
makes the Pandas Series equality tests unsuitable. We use np.allclose so
that we can deal with both the fp precision issues if they arise, and
with the necessary NaN padding as well.